Toujours pas réussi à modifier la durée d'un évènement

+++ 2A_POO_robots2018/src/simulation/EvenementDeplacement.java
@@ -10,7 +10,7 @@ public class EvenementDeplacement extends Evenement{
   Carte carte;
   Robot robot;
 
-  public EvenementDeplacement(int date, Direction dir, Carte carte, Robot robot){
+  public EvenementDeplacement(long date, Direction dir, Carte carte, Robot robot){
     super(date);
     this.dir = dir;
     this.carte = carte;
